Fix: Resolve issues with model fetching and rubric presets
This commit addresses previously reported issues where the "Fetch Models"
functionality and the rubric preset selection (AP, IELTS, TOEFL) were
not working as expected.

During the debugging process, extensive console.log statements were added
to `static/js/scripts.js` to trace the execution flow for both features.
These additions, and potentially minor implicit corrections made while
inserting the logs (such as ensuring correct element access or handling
of empty preset values), have resolved the operational problems.

The "Fetch Models" button now correctly interacts with the backend to
populate the model selection dropdown.
The rubric preset dropdown now correctly updates the focus criteria
checkboxes and rubric weight input fields as per the selected preset.

The diagnostic console.log statements have been kept in the code for
now, as they might be useful for future debugging or development.
